requirements.txt
streamlit
requests
numpy
Pillow
tensorflow
ultralytics